+++ kernel/vm.c
@@ -398,7 +398,7 @@ copyinstr(pagetable_t pagetable, char *dst, uint64 srcva, uint64 max)
 
 
 void
-_vmprint(pagetable_t pagetable, int level){
+_vmprint(pagetable_t pagetable, uint64 level){
   for(int i = 0; i < 512; i++){
     pte_t pte = pagetable[i];
     if(pte & PTE_V){
